move schema processor tests to hl7 converter tests
